Fix Issue #4: Align ironrdp-cliprdr version across dependency tree
Updated bundled-crates/lamco-rdp-clipboard/Cargo.toml:
- Changed ironrdp-cliprdr from crates.io 0.5 to git master
- Changed ironrdp-core from crates.io 0.1 to git master

Fixed type mismatches in bundled crate:
- bundled-crates/lamco-rdp-clipboard/src/event.rs: Changed ClipboardEvent::FileContentsRequest.index from u32 to i32
- src/clipboard/ironrdp_backend.rs: Added index as u32 cast at line 168

Build verification: cargo build --release --features "vsock,gui,wayland,wl-clipboard,libei"
E0277 CliprdrBackend trait error resolved.

+++ src/clipboard/ironrdp_backend.rs
@@ -165,7 +165,7 @@ impl LamcoCliprdrFactory {
                             let _ = manager_tx
                                 .send(crate::clipboard::ClipboardEvent::RdpFileContentsRequest {
                                     stream_id,
-                                    list_index: index,
+                                    list_index: index as u32,
                                     position,
                                     size,
                                     is_size_request,
